Update project structure/rename files/update paths
