+++ src/c/_create_roster.c
@@ -13,9 +13,9 @@ Description : This source file contains the function used
 #include <stdlib.h>
 #include <stdint.h>
 #include <time.h>
-#include "../lib/headers/utils.h"
-#include "../lib/headers/db.hpp"
-#include "../lib/headers/c_files.h"
+
+#include "../lib/headers/cpp/cpp_files.hpp"
+#include "../lib/headers/c/c_files.h"
 
 /************************************************************************************
  * create_new_roster() This is the main function for creating new rosters.
