+++ src/c/_manage_roster.c
@@ -14,9 +14,9 @@ Description : This source file contains all functions used
 #include <stdint.h>
 #include <time.h>
 #include <stdarg.h>
-#include "../lib/headers/utils.h"
-#include "../lib/headers/db.hpp"
-#include "../lib/headers/c_files.h"
+
+#include "../lib/headers/cpp/cpp_files.hpp"
+#include "../lib/headers/c/c_files.h"
 
 Column rosterColumn;
 static uint8_t mainMenuProccess;
