+++ src/c/_manage_student_db.c
@@ -12,9 +12,9 @@ Description : This source file contains the function used
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdint.h>
-#include "../lib/headers/utils.h"
-#include "../lib/headers/db.hpp"
-#include "../lib/headers/c_files.h"
+
+#include "../lib/headers/cpp/cpp_files.hpp"
+#include "../lib/headers/c/c_files.h"
 
 static uint8_t mainMenuProccess;
 int manageStudentDBMenuIsRunning;
