+++ src/c/_search.c
@@ -13,9 +13,9 @@ Description : This source file contains the function used
 #include <stdlib.h>
 #include <stdint.h>
 #include <ctype.h>
-#include "../lib/headers/utils.h"
-#include "../lib/headers/db.hpp"
-#include "../lib/headers/c_files.h"
+
+#include "../lib/headers/cpp/cpp_files.hpp"
+#include "../lib/headers/c/c_files.h"
 
 int searchingForStudent = TRUE;
