+++ src/lib/utils.cpp
@@ -15,7 +15,7 @@ Description : This source file contains several utility functions that are used
 #include <chrono>
 #include <thread>
 #include <ctime>
-#include "headers/utils.hpp"
+#include "headers/cpp/cpp_files.hpp"
 /*
 Allows using elements from
 the 'std' namespace without
